feat(project):project directory refactor
