+++ src/server.js
@@ -2,7 +2,7 @@ import express from "express";
 import React from "react";
 import { renderToString } from "react-dom/server";
 import { StaticRouter } from "react-router-dom/server";
-import App from "./client/App";
+import App from "./App";
 
 const app = express();
