+++ webpack.server.js
@@ -4,7 +4,7 @@ const webpackNodeExternals = require("webpack-node-externals");
 module.exports = {
   mode: "development",
   target: "node",
-  entry: "./src/index.js",
+  entry: "./src/server.js",
   output: {
     filename: "bundle.js",
     path: path.resolve(__dirname, "build"),
